+++ controller/itemsController.js
@@ -7,14 +7,14 @@ module.exports.item_get = async function (req, res) {
   const id = req.params._id;
   try {
     const item = await Item.findById(id).populate("owner");
+    res.render("item/item", {
+      item: item,
+    });
   } catch (e) {
     //can occur CastError: Cast to ObjectId failed for value "create" (type string) at path "_id" for model "item"
     res.send("Something went wrong");
     return;
   }
-  res.render("item/item", {
-    item: item,
-  });
 };
 
 module.exports.item_create_page = async function (req, res) {
