+++ routes/api/authApi.js
@@ -1,0 +1,13 @@
+const express = require("express");
+const router = express.Router();
+const authApiController = require("../../controller/api/authApiController");
+const {
+  requireAuth,
+  checkUser,
+  upload,
+} = require("../../middleware/authMiddleware");
+
+router.post("/signup", authApiController.signup);
+router.post("/login", authApiController.login);
+router.delete("/user/:_id", authApiController.delete_user);
+module.exports = router;
