+++ routes/authRoutes.js
@@ -8,11 +8,8 @@ const {
 } = require("../middleware/authMiddleware");
 
 // user
-router.get("/signup", authController.signup_get);
-router.post("/signup", authController.signup_post);
-router.get("/login", authController.login_get);
-router.post("/login", authController.login_post);
-router.get("/logout", authController.logout_get);
-router.delete("/user/:_id", authController.delete_user);
+router.get("/signup", authController.signup);
+router.get("/login", authController.login);
+router.get("/logout", authController.logout);
 
 module.exports = router;
